Add files via upload
This code creates a basic GUI with a homepage, scenes tab, and input files tab. The scenes tab allows you to select background, character, and text box images, which are then saved to files in the same directory as the app. The undo and export functions are not implemented in this example, but you can add your own implementation for these functions.

Note that this code uses the PIL library to open and save images. You may need to install this library if you haven't already. You can do this by running pip install pillow in your terminal.